Split functionality to more files
