+++ foo_discord_rich/discord/uploader.cpp
@@ -5,6 +5,7 @@
 #include <fb2k/config.h>
 
 #include <ctime>
+#include <fb2k/artwork_metadb.h>
 
 #include "image_hasher.h"
 #include "foobar2000/SDK/component.h"
@@ -103,27 +104,6 @@ void threaded_process_artwork_uploader::on_done(ctx_t p_wnd,bool p_was_aborted)
 {
 }
 
-metadb_index_client_impl::metadb_index_client_impl( const char * pinTo ) {
-    static_api_ptr_t<titleformat_compiler>()->compile_force(m_keyObj, pinTo);
-}
-
-metadb_index_hash metadb_index_client_impl::transform(const file_info & info, const playable_location & location) {
-    pfc::string_formatter str;
-    m_keyObj->run_simple( location, &info, str );
-    // Make MD5 hash of the string, then reduce it to 64-bit metadb_index_hash
-    return static_api_ptr_t<hasher_md5>()->process_single_string( str ).xorHalve();
-}
-
-metadb_index_client_impl * clientByGUID( const GUID & guid ) {
-    // Static instances, never destroyed (deallocated with the process), created first time we get here
-    // Using service_impl_single_t, reference counting disabled
-    // This is somewhat ugly, operating on raw pointers instead of service_ptr, but OK for this purpose
-    static metadb_index_client_impl* g_clientTrack = new service_impl_single_t<metadb_index_client_impl>( config::artworkMetadbKey.GetValue().c_str() );
-
-    PFC_ASSERT( guid == guid::artwork_url_index );
-    return g_clientTrack;
-}
-
 bool extractAndUploadArtwork(const metadb_handle_ptr track, abort_callback &abort, pfc::string8 &artwork_url, metadb_index_hash hash)
 {
     if ( config::uploadArtworkCommand.GetValue().length() == 0 )
@@ -154,7 +134,7 @@ bool extractAndUploadArtwork(const metadb_handle_ptr track, abort_callback &abor
         artwork_url = uploadArtwork( artwork, abort );
         if (artwork_url.get_length() > 0)
         {
-            artwork_url_set( hash, artwork_url );
+            drp::artwork_url_set( hash, artwork_url );
             return true;
         }
     }
@@ -552,294 +532,5 @@ pfc::string8 uploadArtwork(artwork_info& art, abort_callback &abort)
     
     return artwork_url;
 }
-    
-// Static cached ptr to metadb_index_manager
-// Cached because we'll be calling it a lot on per-track basis, let's not pass it everywhere to low level functions
-// Obtaining the pointer from core is reasonably efficient - log(n) to the number of known service classes, but not good enough for something potentially called hundreds of times
-static metadb_index_manager::ptr g_cachedAPI;
-static metadb_index_manager::ptr cached_index_api() {
-    auto ret = g_cachedAPI;
-    if ( ret.is_empty() ) ret = metadb_index_manager::get(); // since fb2k SDK v1.4, core API interfaces have a static get() method
-    return ret;
-}
-    
-void record_set( metadb_index_hash hash, const record_t & record) {
-    stream_writer_formatter_simple< /* using bing endian data? nope */ false > writer;
-    writer << record.artwork_url;
 
-    cached_index_api()->set_user_data( guid::artwork_url_index, hash, writer.m_buffer.get_ptr(), writer.m_buffer.get_size() );
-}
-
-    
-record_t record_get( metadb_index_hash hash) {
-    mem_block_container_impl temp; // this will receive our BLOB
-    cached_index_api()->get_user_data( guid::artwork_url_index, hash, temp );
-    if ( temp.get_size() > 0 ) {
-        try {
-            // Parse the BLOB using stream formatters
-            stream_reader_formatter_simple_ref reader(temp.get_ptr(), temp.get_size());
-
-            record_t ret;
-
-            if ( reader.get_remaining() > 0 ) {
-                reader >> ret.artwork_url;
-            }
-            
-            return ret;
-        } catch (exception_io_data) {
-            // we get here as a result of stream formatter data error
-            // fall thru to return a blank record
-        }
-    }
-    return record_t();
-}
-
-// Our group in Properties dialog / Details tab, see track_property_provider_impl
-static const char strPropertiesGroup[] = "Discord rich presence";
-
-	// An init_stage_callback to hook ourselves into the metadb
-// We need to do this properly early to prevent dispatch_global_refresh() from new fields that we added from hammering playlists etc
-class init_stage_callback_impl : public init_stage_callback
-{
-public:
-    void on_init_stage( t_uint32 stage )
-    {
-        if ( stage == init_stages::after_config_read )
-        {
-            auto api = metadb_index_manager::get();
-            g_cachedAPI = api;
-            // Important, handle the exceptions here!
-            // This will fail if the files holding our data are somehow corrupted.
-            try
-            {
-                api->add( clientByGUID( guid::artwork_url_index ), guid::artwork_url_index, system_time_periods::week * 4 );
-            }
-            catch ( std::exception const& e )
-            {
-                api->remove( guid::artwork_url_index );
-                FB2K_console_formatter() << DRP_NAME_WITH_VERSION << ": Critical initialization failure: " << e;
-                return;
-            }
-            api->dispatch_global_refresh();
-        }
-    }
-};
-class initquit_impl : public initquit
-{
-public:
-    void on_quit()
-    {
-        // Cleanly kill g_cachedAPI before reaching static object destructors or else
-        g_cachedAPI.release();
-    }
-};
-
-static service_factory_single_t<init_stage_callback_impl> g_init_stage_callback_impl;
-static service_factory_single_t<initquit_impl> g_initquit_impl;
-
-    
-bool artwork_url_set( metadb_index_hash hash, const pfc::string8 &artwork_url )
-{
-    auto rec = record_get( hash );
-    bool bChanged = false;
-    if ( ! rec.artwork_url.equals( artwork_url ) ) {
-        rec.artwork_url = artwork_url;
-        record_set( hash, rec );
-        bChanged = true;
-    }
-
-    return bChanged;
-}
-
-    
-
-void generateUrls( metadb_handle_list_cref tracks ) {
-    const size_t count = tracks.get_count();
-    if (count == 0) return;
-
-    auto client = clientByGUID(guid::artwork_url_index);
-    
-    pfc::map_t<metadb_index_hash, metadb_handle_ptr> allHashes;
-    for (size_t w = 0; w < count; ++w) {
-        metadb_index_hash hash;
-        if (client->hashHandle(tracks[w], hash)) {
-            if (allHashes.exists(hash)) continue;
-            allHashes.set(hash, tracks[w]);
-        }
-    }
-
-    if (allHashes.get_count() == 0) {
-        FB2K_console_formatter() << DRP_NAME_WITH_VERSION << ": Could not hash any of the tracks due to unavailable metadata, bailing";
-        return;
-    }
-
-    auto thread_impl = new service_impl_t<threaded_process_artwork_uploader>(allHashes);
-    const std::string p_title = "Uploading artwork";
-
-    threaded_process::g_run_modeless(
-        thread_impl,
-        threaded_process::flag_show_progress | threaded_process::flag_show_abort | threaded_process::flag_show_delayed,
-        g_foobar2000_api->get_main_window(),
-        p_title.c_str(),
-        p_title.length()
-    );
-}
-
-void clearUrls( metadb_handle_list_cref tracks ) {
-    const size_t count = tracks.get_count();
-    if (count == 0) return;
-
-    auto client = clientByGUID(guid::artwork_url_index);
-    pfc::avltree_t<metadb_index_hash> allHashes;
-
-    for (size_t w = 0; w < count; ++w) {
-        metadb_index_hash hash;
-        if (client->hashHandle(tracks[w], hash)) {
-            if (allHashes.exists(hash)) continue;
-            allHashes += hash;
-        }
-    }
-
-    if (allHashes.get_count() == 0) {
-        FB2K_console_formatter() << DRP_NAME_WITH_VERSION << ": Could not hash any of the tracks due to unavailable metadata, bailing";
-        return;
-    }
-
-    pfc::list_t<metadb_index_hash> lstChanged; // Linear list of hashes that actually changed
-
-    for (auto iter = allHashes.first(); iter.is_valid(); ++iter)
-    {
-        const metadb_index_hash hash = *iter;
-        if (artwork_url_set( hash, "" ))
-        {
-            lstChanged += hash;
-        }
-    }
-
-    FB2K_console_formatter() << DRP_NAME_WITH_VERSION << ": " << lstChanged.get_count() << " entries cleared";
-    if (lstChanged.get_count() > 0) {
-        // This gracefully tells everyone about what just changed, in one pass regardless of how many items got altered
-        cached_index_api()->dispatch_refresh(guid::artwork_url_index, lstChanged);
-    }
-}
-    
-class contextmenu_artwork_url : public contextmenu_item_simple {
-public:
-	GUID get_parent() {
-		return guid::context_menu_group;
-	}
-
-	unsigned get_num_items() {
-		return 2;
-	}
-
-	void get_item_name(unsigned p_index, pfc::string_base & p_out) {
-		PFC_ASSERT( p_index < get_num_items() );
-		switch(p_index) {
-		    case 0:
-		        p_out = "Generate artwork url"; break;
-		    case 1:
-		        p_out = "Clear artwork"; break;
-		}
-	}
-
-	void context_command(unsigned p_index, metadb_handle_list_cref p_data, const GUID& p_caller) {
-		PFC_ASSERT( p_index < get_num_items() );
-
-        switch (p_index)
-        {
-        case 0:
-            generateUrls( p_data );
-            break;
-        case 1:
-            clearUrls( p_data );
-            break;
-        default:
-            uBugCheck();    
-        }
-	}
-
-	GUID get_item_guid(unsigned p_index) {
-		switch(p_index) {
-		    case 0:	return guid::context_menu_item_generate_url;
-		    case 1:	return guid::context_menu_item_clear_url;
-		    default: uBugCheck();
-		}
-	}
-
-	bool get_item_description(unsigned p_index, pfc::string_base & p_out) {
-		PFC_ASSERT( p_index < get_num_items() );
-		switch( p_index ) {
-		case 0:
-			p_out = "Generate urls to be used with discord rich presence";
-			return true;
-		case 1:
-		    // Currently clears based on album. Can be changed after a good way of storing image hashes is found
-		    p_out = "Clear artwork urls of selected albums";
-		    return true;
-		default:
-			PFC_ASSERT(!"Should not get here");
-			return false;
-		}
-	}
-};
-
-static contextmenu_group_popup_factory g_mygroup( guid::context_menu_group, contextmenu_groups::root, "Discord rich presence", 0 );
-static contextmenu_item_factory_t< contextmenu_artwork_url > g_contextmenu_rating;
-
-
-class track_property_provider_impl : public track_property_provider_v2 {
-public:
-	void workThisIndex(GUID const & whichID, double priorityBase, metadb_handle_list_cref p_tracks, track_property_callback & p_out) {
-		auto client = clientByGUID( whichID );
-		pfc::avltree_t<metadb_index_hash> hashes;
-		const size_t trackCount = p_tracks.get_count();
-		for (size_t trackWalk = 0; trackWalk < trackCount; ++trackWalk) {
-			metadb_index_hash hash;
-			if (client->hashHandle(p_tracks[trackWalk], hash)) {
-				hashes += hash;
-			}
-		}
-
-		pfc::string8 strComment;
-
-		{
-			bool bFirst = true;
-			bool bVarComments = false;
-			for (auto i = hashes.first(); i.is_valid(); ++i) {
-				auto rec = record_get( *i );
-
-
-				if ( bFirst ) {
-					strComment = rec.artwork_url;
-				} else if ( ! bVarComments ) {
-					if ( strComment != rec.artwork_url ) {
-						bVarComments = true;
-						strComment = "<various>";
-					}
-				}
-
-				bFirst = false;
-			}
-		}
-
-		p_out.set_property(strPropertiesGroup, priorityBase, PFC_string_formatter() << "Artwork Url", strComment);
-	}
-	void enumerate_properties(metadb_handle_list_cref p_tracks, track_property_callback & p_out) {
-		workThisIndex( guid::artwork_url_index, 0, p_tracks, p_out );
-	}
-	void enumerate_properties_v2(metadb_handle_list_cref p_tracks, track_property_callback_v2 & p_out) {
-		if ( p_out.is_group_wanted( strPropertiesGroup ) ) {
-			enumerate_properties( p_tracks, p_out );
-		}
-	}
-	
-	bool is_our_tech_info(const char * p_name) { 
-		// If we do stuff with tech infos read from the file itself (see file_info::info_* methods), signal whether this field belongs to us
-		// We don't do any of this, hence false
-		return false; 
-	}
-};
-
-static service_factory_single_t<track_property_provider_impl> g_track_property_provider_impl;
 }
