+++ foo_discord_rich/discord/uploader.h
@@ -14,21 +14,7 @@ public:
 private:
     pfc::map_t<metadb_index_hash, metadb_handle_ptr> hashes_{};
 };
-    
-// A class that turns metadata + location info into hashes to which our data gets pinned by the backend.
-class metadb_index_client_impl : public metadb_index_client
-{
-public:
-    metadb_index_client_impl( const char * pinTo );
-    metadb_index_hash transform(const file_info & info, const playable_location & location);
-private:
-    titleformat_object::ptr m_keyObj;
-};
 
-/* metadb record */
-struct record_t {
-    pfc::string8 artwork_url;
-};
 
 struct artwork_info
 {
@@ -38,13 +24,8 @@ struct artwork_info
     pfc::string8 artwork_hash;
 };
 
-metadb_index_client_impl * clientByGUID( const GUID & guid );
 bool extractAndUploadArtwork( const metadb_handle_ptr track, abort_callback &abort, pfc::string8 &artwork_url, metadb_index_hash hash );
 artwork_info extractArtwork( const metadb_handle_ptr track, abort_callback &abort );
 pfc::string8 uploadArtwork( artwork_info& art, abort_callback &abort );
-metadb_index_manager::ptr cached_index_api();
-void record_set( metadb_index_hash hash, const record_t & record);
-record_t record_get( metadb_index_hash hash);
-bool artwork_url_set( metadb_index_hash hash, const pfc::string8 &artwork_url );
-    
+
 } // namespace drp::uploader
