+++ foo_discord_rich/foo_discord_rich.vcxproj
@@ -125,7 +125,9 @@
     <ClCompile Include="discord\image_hasher.cpp" />
     <ClCompile Include="discord\uploader.cpp" />
     <ClCompile Include="fb2k\acfu_integration.cpp" />
+    <ClCompile Include="fb2k\artwork_metadb.cpp" />
     <ClCompile Include="fb2k\config.cpp" />
+    <ClCompile Include="fb2k\menus.cpp" />
     <ClCompile Include="fb2k\playback.cpp" />
     <ClCompile Include="dllmain.cpp" />
     <ClCompile Include="fb2k\mainmenu.cpp" />
@@ -142,6 +144,7 @@
     <ClInclude Include="discord\discord_impl.h" />
     <ClInclude Include="discord\image_hasher.h" />
     <ClInclude Include="discord\uploader.h" />
+    <ClInclude Include="fb2k\artwork_metadb.h" />
     <ClInclude Include="fb2k\config.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="stdafx.h" />
