+++ foo_discord_rich/foo_discord_rich.vcxproj.filters
@@ -50,6 +50,12 @@
     <ClCompile Include="discord\image_hasher.cpp">
       <Filter>discord</Filter>
     </ClCompile>
+    <ClCompile Include="fb2k\artwork_metadb.cpp">
+      <Filter>fb2k</Filter>
+    </ClCompile>
+    <ClCompile Include="fb2k\menus.cpp">
+      <Filter>fb2k</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="utils\acfu_github.h">
@@ -84,6 +90,9 @@
     </ClInclude>
     <ClInclude Include="discord\image_hasher.h">
       <Filter>discord</Filter>
+    </ClInclude>
+    <ClInclude Include="fb2k\artwork_metadb.h">
+      <Filter>fb2k</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
